refactored migrations to bigInteger for mobile numbers, removed create.blade, working on history blade

+++ database/migrations/2017_11_27_145100_create_messages_table.php
@@ -18,7 +18,7 @@ class CreateMessagesTable extends Migration
             $table->char('twilio_SID', 34);
             $table->integer('sender_id')->unsigned();
             $table->foreign('sender_id')->references('id')->on('users')->onDelete('cascade');
-            $table->tinyInteger('mobile');
+            $table->bigInteger('mobile');
             $table->string('link', 250);
             $table->datetime('sent_at');
             $table->boolean('is_received')->default(false);
